Relocated purgeEmailFromAllMailboxes() from GreenMailUtil to GreenMail per code review in PR #101 Relocated tests from GreenMailUtilTest to ExamplePurgeAllEmailsTest class Removed related code from GreenMailUtil and GreenMailUtilTest classes Changed access modifier on GreenMailRule.getGreenMail from protected to public
